Change dropdown markdown component to be dynamic
Now the component allows to change the folder where the component is called, in each of the main.mdx files
