--- frontend/src/components/tipDropdown.js
@@ -1,2 +1,0 @@
-import React from 'react';
-
